@@ -1,69 +1,27 @@
-const express = require("express");
-const app = express();
-const session = require("express-session");
-const cookieParser = require("cookie-parser");
-const mongoose = require("mongoose");
-const bodyParser = require("body-parser");
-const cors = require("cors");
-const reactViews = require("express-react-views");
-const fs = require("fs");
-const path = require("path");
-const logger = require("morgan");
-const multer = require("multer");
-const randomstring = require("randomstring");
+import express from "express";
+import path from "path";
+import reactViews from "express-react-views";
+import { fileURLToPath } from "url";
+import configExpress from "./config/express.js";
+import routes from "./routes/index.js";
 
-const authRouter = require("./routes/auth");
-const dashboardRouter = require("./routes/dashboard");
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
-mongoose.Promise = global.Promise;
-mongoose
-  .connect("mongodb://127.0.0.1:27017/studentServices", {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-  })
-  .catch((e) => {
-    console.error("Connection error", e.message);
-  });
-const db = mongoose.connection;
+let app = express();
 
-
-app.use(cookieParser());
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use(cors());
-
-// setup the logger
-app.use(logger("dev"));
+let infoApp = {};
 
 //set static dir
 app.use(express.static(path.join(__dirname, "public")));
 
+// set views path, template engine and default layout
 app.set("views", __dirname + "/views");
 app.set("view engine", "jsx");
 app.engine("jsx", reactViews.createEngine());
+//app.use(favicon(__dirname + '/public/favicon.png'));
 
-app.set("trust proxy", 1);
-app.use(
-  session({
-    secret: randomstring.generate(),
-    resave: false,
-    saveUninitialized: true,
-    cookie: {
-      maxAge: 60000,
-      secure: true,
-      expires: new Date(Date.now() + 60000),
-    },
-  })
-);
+//config
+app = configExpress(app, infoApp);
 
-//routers
-app.get("/", function (req, res) {
-  res.render("home", { name: "student Services", title: "Student Services" });
-});
-app.use("/auth", authRouter);
-app.use("/dashboard", dashboardRouter);
-
-const port = 3000;
-app.listen(port, () => {
-  console.log(`Server running on http://localhost:${port}`);
-});
+// routes
+app.use("/", routes(infoApp));
